refactor + initial phase of taking in a json file and using that to generate multiple schema files at once
+ example schema file we're working with
